move `dir_size` to `common`
